chore: consolidate ucd-store remote test files
